+++ apps/web/app/layouts/default.vue
@@ -15,25 +15,13 @@
     <PreviewMode />
     <QuickCheckout v-if="isOpen" :product="product" />
     <slot name="footer">
-      <PageBlock
-        v-for="(block, index) in footerBlocks"
-        :key="block.meta.uuid"
-        :block="block"
-        :index="index"
-        :root="true"
-        :enable-actions="false"
-        :is-clicked="false"
-        :clicked-block-index="null"
-        :is-tablet="false"
-        :change-block-position="() => {}"
-      />
+      <UiFooterBlocks />
     </slot>
   </div>
 </template>
 
 <script setup lang="ts">
 import type { DefaultLayoutProps } from '~/layouts/types';
-import type { Block } from '@plentymarkets/shop-api';
 
 defineProps<DefaultLayoutProps>();
 
@@ -51,11 +39,6 @@ const useSimplifiedHeader = computed(() => {
     path.includes('/readonly-checkout')
   );
 });
-
-const { globalBlocksCache } = useGlobalBlocks();
-const footerBlocks = computed(
-  () => (globalBlocksCache.value?.filter((block: Block) => block.name === 'Footer') ?? []) as Block[],
-);
 
 setLogoMeta();
 </script>
